Remove dynamic arch (#350)
* Refactor code structure for improved readability and maintainability

* Add NNUE evaluation benchmark tool and documentation

- Implemented a benchmarking tool for NNUE evaluation functions in `bench_nnue_eval.rs`.
- Added command-line interface for specifying NNUE file paths, iterations, warmup counts, and comparison mode.
- Included benchmark tests for various NNUE architectures (HalfKP, HalfKA) with detailed performance metrics.
- Created methodology documentation for NNUE benchmarking in `nnue-benchmark-methodology.md`.
- Updated performance README to reference new benchmark methodology and results.
- Added initial benchmark results in JSON format to `nnue-eval-results.json`.

* fix: evalディレクトリにREADME.mdを追加し、NNUEファイルの配置規約を明確化

* Remove dynamic benchmark functions for HalfKADynamic and HalfKPDynamic from nnue evaluation tool

* refactor: 型パラメータに基づく入力出力の整合性を確認するアサーションを追加し、ベンチマークのデフォルト反復回数を変更

* refactor: SIMD最適化用の定数とメソッドを追加

* Refactor NNUE activation functions to include quantization parameter

- Updated `activate_i32_to_u8` calls in `network_halfka.rs` and `network_halfkp.rs` to pass the quantization parameter `self.qa`.
- Removed the static implementation comparison code from `bench_nnue_eval.rs` to streamline benchmarking.
- Adjusted the benchmark functions to focus on const generics implementations.

* refactor: AccumulatorStackVariantのドキュメントを更新し、ネットワークアーキテクチャのサポートを明確化
refactor: NNUENetworkのアーキテクチャパース処理を改善し、ユニットテストを追加

* refactor: activate_i32_to_u8関数からqa引数を削除し、固定スケーリングを使用するように変更

* feat: Implement const generics based NNUE network

- Unified NNUE evaluation function network implementation using const generics, eliminating code duplication while maintaining and improving performance.
- Added new files for activation functions, network implementations, benchmarking tools, and documentation.
- Removed static implementation of HalfKA, integrating it into the const generics structure.
- Enhanced performance benchmarks showing improvements over static implementations.

feat: Add Threat Inputs implementation task for Shogi

- Initiated investigation, design, and implementation tasks for adapting Full Threat Inputs from Stockfish SFNNv10 to Shogi.
- Documented phases for research, design, implementation, and verification.

chore: Create task notes for ongoing development

- Compiled task notes and references for various ongoing projects related to NNUE and Shogi engine improvements.

feat: Add batch preprocessing and rescoring scripts

- Introduced scripts for batch preprocessing of NNUE data and rescoring using the Shogi engine.
- Implemented logging and error handling for better tracking of processing status.

feat: Add performance profiling scripts

- Developed scripts for generating flamegraphs and analyzing search performance in the Shogi engine.
- Included options for cache statistics and assembly annotations to aid in performance tuning.

* refactor: Update NNUE network architecture handling to support const generics and improve fallback logic

* refactor: Update fallback order for network loading to prioritize larger architecture files

---------

+++ packages/rust-core/crates/engine-core/src/nnue/constants.rs
@@ -155,7 +155,8 @@ pub const SCRELU_QA: i16 = 127;
 /// SCReLU L1層以降の量子化係数 (bullet-shogi 準拠)
 ///
 /// L1層以降では QB = 64 を使用。
-pub const SCRELU_QB: i16 = 64;
+/// i32 として定義（中間層の i32 演算で使用するため）
+pub const SCRELU_QB: i32 = 64;
 
 #[cfg(test)]
 mod tests {
